no longer using Test::Unit and using RSpec
all existing tests have been converted to specs. 

A little note on the TODO in spec/job_spec.rb for some reason in describe block I cannot get the shoulda macros "should_validate_presence_of" and "should_require_readonly_attributes" to work. Spent a bit of time trying to figure this out but havent gotten it working. Since this part of the spec really isn't as important as the behavior of the Job API I'm going to go back and adress it later.