the basics of callbacks (before_invoke)
before_invoke callback allows methods in the Job class to be called right before it is run. Use the macro before_invoke(*do_methods) to specify which methods to call. The callback chain does not interrupt the running of the job. It is arrogant of the results returned by each before_invoke method. Methods are called in the order they are added so
	
	before_invoke :method1, :method2, :method3

is the same as

	before_invoke :method1
	before_invoke :method2
	before_invoke :method3