after_init callback executes given methods after initializing Job
seeing as there are many callbacks to add should probably dynamically call alias_method_chain and define methods like invoke_worker_with_callbacks using cb_points hash. This wont be as simple as the dynamic class method and callers/checkers generation however

Added a "stub" ExampleWorker class to the top of the callbacks_spec. This way background-fu files need not be generated to run specs. Had to make some changes in places where the use of ExampleWorker#add was not stubbed.