simple magic (worker) callbacks
this is the beginning of work on magic callbacks or worker callbacks. In the first use case (working as of this commit) you simply want to call one method before and after the method your worker runs when called by a Job. Say you have a worker, ExampleWorker with a method add and you enqueue a job with, Job.enqueue!(ExampleWorker, :add, 1, 2). When the Job is run if the instance methods before_add or after_add are defined in Example worker they will be called in the logical places. Use these magic callbacks to perform tasks such as logging or modifying the state of an object prior to using it in a long running Job. This first use case is really a simplification of the second.

The second use case (not currently implemented) is wanting to call multiple methods before the worker method. This will be possible in similar fashion to Job callbacks. Extending the example above it will be possible to use the class macros before_add/after_add passing them a list of method names to call at the appropriate times.  

Also looking into adding hooks like rescue_add and ensure_add.