Job run chain callbacks complete
added before_ensure and after_ensure. these work in similar fashion to before_invoke and other callbacks. Methods registered with before_ensure or after_ensure are guarunteed to be called if the worker errors. If you opened a file for some reason in before_invoke then you should close the file in a before_ensure or after_ensure based on your needs.