added necessary requires to Railtie
Rails 3 requires developers to explicitly require any pieces of Rails they use as well as including the lib itselft. This is why b-fu and rails are required in the Railtie. rails/engine is required because the Railtie hooks into set_load_path which is an initializer defined in rails/engine.