Background-Fu config file is now config/background.rb instead of config/daemons.rb. No longer use the rails initialization process load workers.
The workers are just required if included in lib/workers. There is no reason to add the the load path when lib dir is already in it. Also, this will automatically require workers in the background script.  

The Railtie at this point is empty and could be done away with but since we will be using Rails::Engine soon I have just left the file.