no longer using daemons gem.
instead of using the old: 
  
    ./script/daemons start

use:

    ruby script/background

If you refer the ./script method just make script/background executable.

Currently, script/background is missing some of its original features. For examle it logs to the console when run as opposed to the Rails logs. The original functionality will be restored. The script now recognizes the file config/background.yml as its config file.